Add Saraga Carnatic and Hindustani loaders
