+++ scripts/make_saraga_hindustani_index.py
@@ -1,0 +1,117 @@
+import argparse
+import hashlib
+import json
+import os
+
+
+SARAGA_HINDUSTANI_INDEX_PATH = '../mirdata/datasets/indexes/saraga_hindustani_index.json'
+
+
+def md5(file_path):
+    """Get md5 hash of a file.
+
+    Parameters
+    ----------
+    file_path: str
+        File path.
+
+    Returns
+    -------
+    md5_hash: str
+        md5 hash of data in file_path
+    """
+    hash_md5 = hashlib.md5()
+    with open(file_path, 'rb') as fhandle:
+        for chunk in iter(lambda: fhandle.read(4096), b''):
+            hash_md5.update(chunk)
+    return hash_md5.hexdigest()
+
+
+def make_saraga_hindustani_index(dataset_data_path):
+
+    saraga_index = {
+        'version': 1.5,
+        'tracks': {},
+    }
+    idx = 0
+    for concert in os.listdir(dataset_data_path):
+        if '.' not in concert:
+            for song in os.listdir(os.path.join(dataset_data_path, concert)):
+                if '.' not in song:
+
+                    # Declare track attributes
+                    index = str(idx) + '_' + song.replace(' ', '_')
+                    print(index)
+                    audio = (None, None)
+                    ctonic = (None, None)
+                    pitch = (None, None)
+                    tempo = (None, None)
+                    sama = (None, None)
+                    sections = (None, None)
+                    phrases = (None, None)
+                    metadata = (None, None)
+
+                    for file in os.listdir(os.path.join(dataset_data_path, concert, song)):
+                        if '.mp3' in file:
+                            audio_path = os.path.join(concert, song, file)
+                            audio_checksum = md5(os.path.join(dataset_data_path, audio_path))
+                            audio = (audio_path, audio_checksum)
+                        if 'tonic' in file:
+                            ctonic_path = os.path.join(concert, song, file)
+                            ctonic_checksum = md5(os.path.join(dataset_data_path, ctonic_path))
+                            ctonic = (ctonic_path, ctonic_checksum)
+                        if 'pitch.' in file:
+                            pitch_path = os.path.join(concert, song, file)
+                            pitch_checksum = md5(os.path.join(dataset_data_path, pitch_path))
+                            pitch = (pitch_path, pitch_checksum)
+                        if 'tempo' in file:
+                            tempo_path = os.path.join(concert, song, file)
+                            tempo_checksum = md5(os.path.join(dataset_data_path, tempo_path))
+                            tempo = (tempo_path, tempo_checksum)
+                        if 'sama' in file:
+                            sama_path = os.path.join(concert, song, file)
+                            sama_checksum = md5(os.path.join(dataset_data_path, sama_path))
+                            sama = (sama_path, sama_checksum)
+                        if 'sections-manual-p' in file:
+                            sections_path = os.path.join(concert, song, file)
+                            sections_checksum = md5(os.path.join(dataset_data_path, sections_path))
+                            sections = (sections_path, sections_checksum)
+                        if 'phrase' in file:
+                            phrases_path = os.path.join(concert, song, file)
+                            phrases_checksum = md5(os.path.join(dataset_data_path, phrases_path))
+                            phrases = (phrases_path, phrases_checksum)
+                        if '.json' in file:
+                            metadata_path = os.path.join(concert, song, file)
+                            metadata_checksum = md5(os.path.join(dataset_data_path, metadata_path))
+                            metadata = (metadata_path, metadata_checksum)
+
+                        saraga_index['tracks'][index] = {
+                            'audio': audio,
+                            'ctonic': ctonic,
+                            'pitch': pitch,
+                            'tempo': tempo,
+                            'sama': sama,
+                            'sections': sections,
+                            'phrases': phrases,
+                            'metadata': metadata
+                        }
+
+                    idx = idx + 1
+
+    with open(SARAGA_HINDUSTANI_INDEX_PATH, 'w') as fhandle:
+        json.dump(saraga_index, fhandle, indent=2)
+
+
+def main(args):
+    print("creating index...")
+    make_saraga_hindustani_index(args.dataset_data_path)
+    print("done!")
+
+
+if __name__ == '__main__':
+    PARSER = argparse.ArgumentParser(description='Make Saraga Hindustani index file.')
+    PARSER.add_argument(
+        'dataset_data_path', type=str, help='Path to Saraga Hindustani data folder.'
+    )
+
+    main(PARSER.parse_args())
