--- tests/resources/mir_datasets/saraga/saraga1.0/carnatic/1/Cherthala Ranganatha Sharma - Bhuvini Dasudane.bpm-manual.txt
@@ -1,3 +1,0 @@
-82,0.065306122,85.35510204
-82,85.35510204,167.314285714
-82,167.314285714,309.342040816
