--- tests/resources/mir_datasets/saraga/saraga1.0/carnatic/1/Cherthala Ranganatha Sharma - Bhuvini Dasudane.ctonic.txt
@@ -1,1 +1,0 @@
-201.740890
