--- tests/resources/mir_datasets/saraga/saraga1.0/carnatic/1/Cherthala Ranganatha Sharma - Bhuvini Dasudane.json
@@ -1,1 +1,0 @@
-{"raaga": [{"uuid": "42dd0ccb-f92a-4622-ae5d-a3be571b4939", "name": "\u015ar\u012branjani"}], "form": [{"name": "Kriti"}], "title": "Bhuvini Dasudane", "work": [{"mbid": "4d05ce9b-c45e-4c85-9eca-941d68b61132", "title": "Bhuvini Dasudane"}], "length": 309000, "taala": [{"uuid": "c788c38a-b53a-48cb-b7bf-d11769260c4d", "name": "\u0100di"}], "album_artists": [{"mbid": "e09b0542-84e1-45ad-b09a-a05a9ad0cb83", "name": "Cherthala Ranganatha Sharma"}], "mbid": "9f5a5452-14cb-4af0-9289-4833854ee60d", "artists": [{"instrument": {"mbid": "c5aa7d98-c14d-4ff1-8afb-f8743c62496c", "name": "Ghatam"}, "attributes": "", "lead": false, "artist": {"mbid": "19f93366-5d58-47f1-bc4f-9225ac7af6ba", "name": "N Guruprasad"}}, {"instrument": {"mbid": "f689271c-37bc-4c49-92a3-a14b15ee5d0e", "name": "Mridangam"}, "attributes": "", "lead": false, "artist": {"mbid": "39c1d741-6154-418b-bf4b-12c77ba13873", "name": "Srimushnam V Raja Rao"}}, {"instrument": {"mbid": "089f123c-0f7d-4105-a64e-49de81ca8fa4", "name": "Violin"}, "attributes": "", "lead": false, "artist": {"mbid": "a2df55e3-d141-4767-862e-77adca691d4b", "name": "B.U. Ganesh Prasad"}}, {"instrument": {"mbid": "d92884b7-ee0c-46d5-96f3-918196ba8c5b", "name": "Voice"}, "attributes": "lead vocals", "lead": true, "artist": {"mbid": "e09b0542-84e1-45ad-b09a-a05a9ad0cb83", "name": "Cherthala Ranganatha Sharma"}}], "concert": [{"mbid": "0816586d-c83e-4c79-a0aa-9b0e578f408d", "title": "Cherthala Ranganatha Sharma at Arkay"}]}
