--- tests/resources/mir_datasets/saraga/saraga1.0/carnatic/1/Cherthala Ranganatha Sharma - Bhuvini Dasudane.mphrases-manual.txt
@@ -1,3 +1,0 @@
-0.224489795	0	2.493877551	ndmdnsndn
-5.844897959	0	2.473469387	ndmdnsndn
-8.504308390	0	2.275555555	ndmdndmgr
