--- tests/resources/mir_datasets/saraga/saraga1.0/carnatic/1/Cherthala Ranganatha Sharma - Bhuvini Dasudane.pitch.txt
@@ -1,6 +1,0 @@
-0.0000000	0.0000000
-0.0044444	100.1200000
-0.0088889	200.2300000
-0.0133333	300.3400000
-0.0177778	400.4300000
-0.0222222	600.12300000
