--- tests/resources/mir_datasets/saraga/saraga1.0/carnatic/1/Cherthala Ranganatha Sharma - Bhuvini Dasudane.sama-manual.txt
@@ -1,3 +1,0 @@
-4.894
-10.229
-15.724
