--- tests/resources/mir_datasets/saraga/saraga1.0/carnatic/1/Cherthala Ranganatha Sharma - Bhuvini Dasudane.sections-manual-p.txt
@@ -1,3 +1,0 @@
-0.065306122	1	85.289795918	Pallavi
-85.35510204	1	81.959183674	Anupallavi
-167.314285714	1	142.027755102	Caraṇam
