--- tests/resources/mir_datasets/saraga/saraga1.0/carnatic/1/Cherthala Ranganatha Sharma - Bhuvini Dasudane.sections-manual.txt
@@ -1,3 +1,0 @@
-0.065306122	1	85.224489795	pallavi
-85.355102040	1	81.502040816	anupallavi
-167.314285714	1	141.453061224	charanam
