--- tests/resources/mir_datasets/saraga/saraga1.0/carnatic/1/Cherthala Ranganatha Sharma - Bhuvini Dasudane.tempo-manual.txt
@@ -1,1 +1,0 @@
-330, 82, 5.827, 32, 4
