--- tests/resources/mir_datasets/saraga/saraga1.0/hindustani/1/Ajoy Chakrabarty - Bairagi.bpm-manual.txt
@@ -1,3 +1,0 @@
--,3.298,59.49
-13,59.49,679.834
-185,679.834,898.664489796
