--- tests/resources/mir_datasets/saraga/saraga1.0/hindustani/1/Ajoy Chakrabarty - Bairagi.ctonic.txt
@@ -1,1 +1,0 @@
-138.591315
