--- tests/resources/mir_datasets/saraga/saraga1.0/hindustani/1/Ajoy Chakrabarty - Bairagi.json
@@ -1,1 +1,0 @@
-{"title": "Bairagi", "raags": [{"common_name": "Bairagi", "uuid": "b143adaa-f1a6-4de4-8985-a5bd35e96279", "name": "Bair\u0101gi"}], "length": 899469, "album_artists": [{"mbid": "653fa2f8-85f8-4829-871f-7c2506ea9b48", "name": "Ajoy Chakrabarty"}], "forms": [{"common_name": "Khayal", "uuid": "7ed81b92-aea6-4f4b-bffb-c12d80012d37", "name": "Khy\u0101l"}], "mbid": "b71c2774-2532-4692-8761-5452e2a83118", "artists": [{"instrument": {"mbid": "d92884b7-ee0c-46d5-96f3-918196ba8c5b", "name": "Voice"}, "attributes": "lead vocals", "lead": true, "artist": {"mbid": "653fa2f8-85f8-4829-871f-7c2506ea9b48", "name": "Ajoy Chakrabarty"}}, {"instrument": {"mbid": "c43c7647-077d-4d60-a01b-769de71b82f2", "name": "Harmonium"}, "attributes": "", "lead": false, "artist": {"mbid": "afbb34e8-1f87-4dd4-81ec-b6145af4d72f", "name": "Paromita Mukherjee"}}, {"instrument": {"mbid": "18e6998b-e53b-415b-b484-d3ac286da99d", "name": "Tabla"}, "attributes": "", "lead": false, "artist": {"mbid": "beee80e6-aa99-451c-9edb-dcda8c2fce8a", "name": "Indranil Bhaduri"}}], "release": [{"mbid": "ae0f2366-9a4f-4534-9376-ac123e881f64", "title": "Geetinandan : Part-3"}], "works": [{"mbid": "b8925ff6-9c8f-4184-8fc8-d358cfdea79b", "title": "Mere Maname Baso Ram Abhiram Puran Ho Sab Kaam"}, {"mbid": "d7a184c3-0187-4912-8708-8d12a4bd9b0a", "title": "Bar Bar Har Gai"}], "taals": [{"common_name": "Ektaal", "uuid": "7cb20903-5f64-4f15-8713-2fb4fcca2b5b", "name": "\u0113kt\u0101l"}, {"common_name": "Ektaal", "uuid": "7cb20903-5f64-4f15-8713-2fb4fcca2b5b", "name": "\u0113kt\u0101l"}], "layas": [{"common_name": "Vilambit", "uuid": "ee58d24a-60aa-4b16-bfcf-edd105118738", "name": "Vila\u1e41bit"}]}
