--- tests/resources/mir_datasets/saraga/saraga1.0/hindustani/1/Ajoy Chakrabarty - Bairagi.mphrases-manual.txt
@@ -1,45 +1,0 @@
-3.506213151	0	7.383945578	Pmr
-12.538775510	2	6.385487528	PnS
-23.382494331	1	8.870022675	rmP
-36.664308390	2	5.549569160	Pmr
-47.415147392	0	6.965986394	PnS
-68.452426303	0	4.295691609	PnS
-74.118095238	1	9.427301587	nrS
-93.878276643	0	7.244625850	rmP
-110.341224489	0	3.459773242	mrS
-114.544036281	1	2.948934240	Srm
-131.215963718	0	9.589841269	PnS
-153.089160997	1	7.546485260	Pnr
-169.668208616	1	3.900952380	rmPn
-189.242630385	0	4.133151927	Srm
-198.460952380	0	3.831292517	Pmr
-203.383582766	0	4.667210884	rmP
-214.134421768	0	4.109931972	Pmr
-241.859047619	0	7.523265306	rmP
-251.634648526	1	4.458231292	PmrS
-257.416417233	0	4.179591836	rmP
-271.673469387	0	3.157913832	Pmr
-297.192199546	0	4.109931972	rmPn
-304.994104308	0	2.020136054	rmPn
-308.500317460	1	3.900952380	mPn
-313.701587301	0	1.602176870	Pmr
-328.005079365	0	7.569705215	PnS
-344.584126984	0	3.970612244	PnS
-351.132154195	0	3.041814058	PnS
-356.124444444	0	3.366893424	PnS
-362.533151927	0	2.345215419	Pmr
-376.116825396	0	6.153287981	PnS
-402.448253968	0	2.577414965	PnS
-407.510204081	0	4.527891156	mrS
-434.612244897	0	3.591836734	rmP
-461.775238095	0	5.294149659	PnS
-488.710385487	0	4.086712018	mrS
-530.253061224	0	3.559183673	mPn
-534.857142857	0	3.330612244	PnS
-711.673469387	0	3.134693877	PnS
-722.253061224	0	3.167346938	PnS
-784.021768707	0	6.548027210	PnS
-794.644897959	0	2.644897959	PnS
-798.432653061	0	2.546938775	PnS
-802.318367346	0	2.351020408	PnS
-889.404081632	0	3.493877551	PnS
