--- tests/resources/mir_datasets/saraga/saraga1.0/hindustani/1/Ajoy Chakrabarty - Bairagi.sama-manual.txt
@@ -1,64 +1,0 @@
-68.385
-123.804
-179.069
-234.339
-289.608
-345.188
-400.433
-455.927
-511.098
-565.800
-620.682
-674.139
-684.280
-688.322
-692.262
-696.215
-700.167
-704.070
-707.965
-711.895
-715.831
-719.726
-723.661
-727.562
-731.498
-735.383
-739.283
-743.160
-747.056
-750.822
-754.712
-758.559
-762.421
-766.287
-770.177
-774.002
-777.892
-781.735
-785.676
-789.548
-793.449
-797.282
-801.151
-805.012
-808.927
-812.850
-816.812
-820.616
-824.502
-828.339
-832.110
-835.865
-839.710
-843.540
-847.319
-851.190
-854.918
-858.682
-862.502
-866.310
-870.207
-874.012
-877.898
-881.682
