--- tests/resources/mir_datasets/saraga/saraga1.0/hindustani/1/Ajoy Chakrabarty - Bairagi.sections-manual-p.txt
@@ -1,3 +1,0 @@
-3.298,1,56.192,Ālāp
-59.49,2,620.344,Khyāl (vilambit ēktāl)
-679.834,3,218.830489796,Khyāl (dr̥t ēktāl)
