--- tests/resources/mir_datasets/saraga/saraga1.0/hindustani/1/Ajoy Chakrabarty - Bairagi.sections-manual.txt
@@ -1,3 +1,0 @@
-3.298,1,54.938,alap
-59.490,2,618.519,vilambit_Ektal
-679.834,3,214.599,drut_Ektal
