+++ tests/resources/mir_datasets/saraga_carnatic/Cherthala Ranganatha Sharma at Arkay by Cherthala Ranganatha Sharma/Bhuvini Dasudane/Bhuvini Dasudane.bpm-manual.txt
@@ -1,0 +1,3 @@
+82,0.065306122,85.35510204
+82,85.35510204,167.314285714
+82,167.314285714,309.342040816
