+++ tests/resources/mir_datasets/saraga_carnatic/Cherthala Ranganatha Sharma at Arkay by Cherthala Ranganatha Sharma/Bhuvini Dasudane/Bhuvini Dasudane.json
@@ -1,0 +1,1 @@
+{"mbid": "9f5a5452-14cb-4af0-9289-4833854ee60d", "title": "Bhuvini Dasudane", "length": 309000, "artists": [{"artist": {"mbid": "19f93366-5d58-47f1-bc4f-9225ac7af6ba", "name": "N Guruprasad"}, "instrument": {"mbid": "c5aa7d98-c14d-4ff1-8afb-f8743c62496c", "name": "Ghatam"}, "lead": false, "attributes": ""}, {"artist": {"mbid": "39c1d741-6154-418b-bf4b-12c77ba13873", "name": "Srimushnam V Raja Rao"}, "instrument": {"mbid": "f689271c-37bc-4c49-92a3-a14b15ee5d0e", "name": "Mridangam"}, "lead": false, "attributes": ""}, {"artist": {"mbid": "a2df55e3-d141-4767-862e-77adca691d4b", "name": "B.U. Ganesh Prasad"}, "instrument": {"mbid": "089f123c-0f7d-4105-a64e-49de81ca8fa4", "name": "Violin"}, "lead": false, "attributes": ""}, {"artist": {"mbid": "e09b0542-84e1-45ad-b09a-a05a9ad0cb83", "name": "Cherthala Ranganatha Sharma"}, "instrument": {"mbid": "d92884b7-ee0c-46d5-96f3-918196ba8c5b", "name": "Voice"}, "lead": true, "attributes": "lead vocals"}], "raaga": [{"uuid": "42dd0ccb-f92a-4622-ae5d-a3be571b4939", "name": "\u015ar\u012branjani", "common_name": "shri ranjani"}], "taala": [{"uuid": "c788c38a-b53a-48cb-b7bf-d11769260c4d", "name": "\u0100di", "common_name": "adi"}], "form": [{"name": "Kriti"}], "work": [{"mbid": "4d05ce9b-c45e-4c85-9eca-941d68b61132", "title": "Bhuvini Dasudane"}], "concert": [{"mbid": "0816586d-c83e-4c79-a0aa-9b0e578f408d", "title": "Cherthala Ranganatha Sharma at Arkay"}], "album_artists": [{"mbid": "e09b0542-84e1-45ad-b09a-a05a9ad0cb83", "name": "Cherthala Ranganatha Sharma"}]}
