+++ tests/resources/mir_datasets/saraga_carnatic/Cherthala Ranganatha Sharma at Arkay by Cherthala Ranganatha Sharma/Bhuvini Dasudane/Bhuvini Dasudane.pitch-vocal.txt
@@ -1,0 +1,6 @@
+0.000000000000000000e+00	0.000000000000000000e+00
+2.902494331065759697e-03	1.123456789012345678e+02
+5.804988662131519393e-03	2.234567890123456789e+02
+8.707482993197278656e-03	3.345678901234567890e+02
+1.160997732426303879e-02	4.456789012345678901e+01
+1.451247165532879892e-02	0.000000000000000000e+00
