+++ tests/resources/mir_datasets/saraga_hindustani/Geetinandan : Part-3 by Ajoy Chakrabarty/Bairagi/Bairagi.bpm-manual.txt
@@ -1,0 +1,3 @@
+-,3.298,59.49
+13,59.49,679.834
+185,679.834,898.664489796
