+++ tests/resources/mir_datasets/saraga_hindustani/Geetinandan : Part-3 by Ajoy Chakrabarty/Bairagi/Bairagi.json
@@ -1,0 +1,1 @@
+{"mbid": "b71c2774-2532-4692-8761-5452e2a83118", "title": "Bairagi", "length": 899469, "artists": [{"artist": {"mbid": "653fa2f8-85f8-4829-871f-7c2506ea9b48", "name": "Ajoy Chakrabarty"}, "instrument": {"mbid": "d92884b7-ee0c-46d5-96f3-918196ba8c5b", "name": "Voice"}, "lead": true, "attributes": "lead vocals"}, {"artist": {"mbid": "afbb34e8-1f87-4dd4-81ec-b6145af4d72f", "name": "Paromita Mukherjee"}, "instrument": {"mbid": "c43c7647-077d-4d60-a01b-769de71b82f2", "name": "Harmonium"}, "lead": false, "attributes": ""}, {"artist": {"mbid": "beee80e6-aa99-451c-9edb-dcda8c2fce8a", "name": "Indranil Bhaduri"}, "instrument": {"mbid": "18e6998b-e53b-415b-b484-d3ac286da99d", "name": "Tabla"}, "lead": false, "attributes": ""}], "raags": [{"uuid": "b143adaa-f1a6-4de4-8985-a5bd35e96279", "common_name": "Bairagi", "name": "Bair\u0101gi"}], "taals": [{"uuid": "7cb20903-5f64-4f15-8713-2fb4fcca2b5b", "common_name": "Ektaal", "name": "\u0113kt\u0101l"}, {"uuid": "7cb20903-5f64-4f15-8713-2fb4fcca2b5b", "common_name": "Ektaal", "name": "\u0113kt\u0101l"}], "layas": [{"uuid": "ee58d24a-60aa-4b16-bfcf-edd105118738", "common_name": "Vilambit", "name": "Vila\u1e41bit"}], "forms": [{"uuid": "7ed81b92-aea6-4f4b-bffb-c12d80012d37", "common_name": "Khayal", "name": "Khy\u0101l"}], "works": [{"mbid": "b8925ff6-9c8f-4184-8fc8-d358cfdea79b", "title": "Mere Maname Baso Ram Abhiram Puran Ho Sab Kaam"}, {"mbid": "d7a184c3-0187-4912-8708-8d12a4bd9b0a", "title": "Bar Bar Har Gai"}], "release": [{"mbid": "ae0f2366-9a4f-4534-9376-ac123e881f64", "title": "Geetinandan : Part-3"}], "album_artists": [{"mbid": "653fa2f8-85f8-4829-871f-7c2506ea9b48", "name": "Ajoy Chakrabarty"}]}
