+++ tests/resources/mir_datasets/saraga_hindustani/Geetinandan : Part-3 by Ajoy Chakrabarty/Bairagi/Bairagi.pitch.txt
@@ -1,0 +1,6 @@
+0.000000000000000000e+00	0.000000000000000000e+00
+4.444444444444444441e-03	1.111111111111111111e+01
+8.888888888888888881e-03	2.222222222222222222e+02
+1.333333333333333419e-02	3.333333333333333333e+02
+1.777777777777777776e-02	4.444444444444444444e+02
+2.222222222222222307e-02	0.000000000000000000e+00
