+++ tests/resources/mir_datasets/saraga_hindustani/Geetinandan : Part-3 by Ajoy Chakrabarty/Bairagi/Bairagi.sections-manual-p.txt
@@ -1,0 +1,3 @@
+3.298,1,56.192,Ālāp
+59.49,2,620.344,Khyāl (vilambit ēktāl)
+679.834,3,218.830489796,Khyāl (dr̥t ēktāl)
