+++ tests/resources/mir_datasets/saraga_hindustani/Geetinandan : Part-3 by Ajoy Chakrabarty/Bairagi/Bairagi.tempo-manual.txt
@@ -1,0 +1,3 @@
+-1, -1, -1, -1, 3.298, 58.236
+13, 4.605, 55.265, 12, 59.490, 678.009
+185, 0.324, 3.885, 12, 679.834, 894.433
