+++ tests/test_loaders.py
@@ -30,7 +30,8 @@ CUSTOM_TEST_TRACKS = {
     "rwc_jazz": "RM-J004",
     "rwc_popular": "RM-P001",
     "salami": "2",
-    "saraga": "carnatic_1",
+    "saraga_carnatic": "126_Bhuvini_Dasudane",
+    "saraga_hindustani": "59_Bairagi",
     "tinysol": "Fl-ord-C4-mf-N-T14d",
 }
